utils: adds ssh client prototype
* Addresses #71
